Add files via upload
I exported the CSV file from my netflix account and had a spreadsheet file that contained two columns, one for Title and the second for the Watch Date. From there I created three more columns to separate the seasons and episodes from the titles in the first column and one for the genre of each show. After I cleaned and prepared the data I ended up with a spreadsheet file of five columns, Title, Season, Episode, Genre, and Watch Date, removing movies and partially viewed titles. In this file I only included shows I’ve watched at least the full first episode of and on. I then created a pivot table illustrating the most viewed genres in the collected data.